+++ src/datastore/interface.go
@@ -11,6 +11,7 @@ import (
 type Datastore interface {
 	ExecuteQuery(user common.User, database string, query *parser.Query, yield func(*protocol.Series) error) error
 	WriteSeriesData(database string, series *protocol.Series) error
+	DropDatabase(database string) error
 	DeleteRangeOfSeries(database, series string, startTime, endTime time.Time) error
 	DeleteRangeOfRegex(user common.User, database string, regex *regexp.Regexp, startTime, endTime time.Time) error
 	Close()